+++ odoo/local-src/specific_stock/models/product_category.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+# © 2016 Julien Coux (Camptocamp)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+
+from openerp import models, fields
+
+
+class ProductCategory(models.Model):
+    _inherit = 'product.category'
+
+    life_time = fields.Integer(
+        string='Product Life Time',
+        help='When a new a Serial Number is issued, '
+             'this is the number of days before the goods '
+             'may become dangerous and must not be consumed.'
+    )
+
+    use_time = fields.Integer(
+        string='Product Use Time',
+        help='When a new a Serial Number is issued, '
+             'this is the number of days before the goods '
+             'starts deteriorating, without being dangerous yet.'
+    )
+
+    removal_time = fields.Integer(
+        string='Product Removal Time',
+        help='When a new a Serial Number is issued, '
+             'this is the number of days before the goods '
+             'should be removed from the stock.'
+    )
+
+    alert_time = fields.Integer(
+        string='Product Alert Time',
+        help='When a new a Serial Number is issued, '
+             'this is the number of days before an alert should be notified.'
+    )