+++ HISTORY.rst
@@ -12,6 +12,8 @@ Unreleased
 
 * Change admin password at the end of setup
 * ALCN-623: Manage stock quants expiration
+* ALCN-624: Manage stock production lot expired dates
+* ALCN-677: Specific manage stock production lot expired dates
 
 **Bugfixes**